[Customer Center] Migrate to List style (#4190)
- Migrates our custom Customer Center UI to use Lists instead of stacks
of buttons
- Implements a reusable navigation stack component with backwards
compatibility
- Pulls out SubscriptionDetailsView into a separate file

**ManageSubscriptionsView**

| New | Old |
|-----|-----|
|
![](https://github.com/user-attachments/assets/5ff13f20-db5d-4e04-a580-fd50b4c47c15)
|
![](https://github.com/user-attachments/assets/618293ce-40cd-4b1b-8945-eeb2844af912)
|


**FeedbackSurveyView**

| New | Old |
|-----|-----|
|
![](https://github.com/user-attachments/assets/828bf86c-8b9e-462f-90ef-cb3da1414549)
|
![](https://github.com/user-attachments/assets/e2ae947e-594a-4d01-8fdb-c57c3963f38b)
|
