+++ RevenueCatUI/CustomerCenter/Views/FeedbackSurveyView.swift
@@ -20,7 +20,7 @@ import SwiftUI
 
 #if os(iOS)
 
-@available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
+@available(iOS 15.0, *)
 @available(macOS, unavailable)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
@@ -46,9 +46,7 @@ struct FeedbackSurveyView: View {
                 background.edgesIgnoringSafeArea(.all)
             }
 
-            VStack {
-                Spacer()
-
+            List {
                 FeedbackSurveyButtonsView(options: self.viewModel.feedbackSurveyData.configuration.options,
                                           onOptionSelected: self.viewModel.handleAction(for:),
                                           loadingState: self.$viewModel.loadingState)
@@ -68,7 +66,7 @@ struct FeedbackSurveyView: View {
 
 }
 
-@available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
+@available(iOS 15.0, *)
 @available(macOS, unavailable)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
@@ -83,33 +81,20 @@ struct FeedbackSurveyButtonsView: View {
     var loadingState: String?
 
     var body: some View {
-        VStack(spacing: Self.buttonSpacing) {
-            ForEach(options, id: \.id) { option in
-                AsyncButton(action: {
-                    await self.onOptionSelected(option)
-                }, label: {
-                    if self.loadingState == option.id {
-                        TintedProgressView()
-                    } else {
-                        Text(option.title)
-                    }
-                })
-                .buttonStyle(ProminentButtonStyle())
-                .disabled(self.loadingState != nil)
+        ForEach(options, id: \.id) { option in
+            AsyncButton {
+                await self.onOptionSelected(option)
+            } label: {
+                if self.loadingState == option.id {
+                    TintedProgressView()
+                } else {
+                    Text(option.title)
+                }
             }
+            .disabled(self.loadingState != nil)
         }
-        .padding([.horizontal, .bottom])
+
     }
-
-}
-
-@available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
-@available(macOS, unavailable)
-@available(tvOS, unavailable)
-@available(watchOS, unavailable)
-extension FeedbackSurveyButtonsView {
-
-    private static let buttonSpacing: CGFloat = 16
 
 }
