+++ RevenueCatUI/CustomerCenter/Views/ManageSubscriptionsView.swift
@@ -20,7 +20,7 @@ import SwiftUI
 
 #if os(iOS)
 
-@available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
+@available(iOS 15.0, *)
 @available(macOS, unavailable)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
@@ -51,29 +51,23 @@ struct ManageSubscriptionsView: View {
     }
 
     var body: some View {
-        let accentColor = Color.from(colorInformation: self.appearance.accentColor, for: self.colorScheme)
-
         if #available(iOS 16.0, *) {
-            NavigationStack {
-                content
-                    .navigationDestination(isPresented: .isNotNil(self.$viewModel.feedbackSurveyData)) {
-                        if let feedbackSurveyData = self.viewModel.feedbackSurveyData {
-                            FeedbackSurveyView(feedbackSurveyData: feedbackSurveyData)
-                        }
+            content
+                .navigationDestination(isPresented: .isNotNil(self.$viewModel.feedbackSurveyData)) {
+                    if let feedbackSurveyData = self.viewModel.feedbackSurveyData {
+                        FeedbackSurveyView(feedbackSurveyData: feedbackSurveyData)
                     }
-            }.applyIf(accentColor != nil, apply: { $0.tint(accentColor) })
+                }
         } else {
-            NavigationView {
-                content
-                    .background(NavigationLink(
-                        destination: self.viewModel.feedbackSurveyData.map { data in
-                            FeedbackSurveyView(feedbackSurveyData: data)
-                        },
-                        isActive: .isNotNil(self.$viewModel.feedbackSurveyData)
-                    ) {
-                        EmptyView()
-                    })
-            }.applyIf(accentColor != nil, apply: { $0.tint(accentColor) })
+            content
+                .background(NavigationLink(
+                    destination: self.viewModel.feedbackSurveyData.map { data in
+                        FeedbackSurveyView(feedbackSurveyData: data)
+                    },
+                    isActive: .isNotNil(self.$viewModel.feedbackSurveyData)
+                ) {
+                    EmptyView()
+                })
         }
     }
 
@@ -84,27 +78,30 @@ struct ManageSubscriptionsView: View {
                 background.edgesIgnoringSafeArea(.all)
             }
 
-            ScrollView {
-                VStack {
-                    if self.viewModel.isLoaded {
-                        SubtitleTextView(subtitle: self.viewModel.screen.subtitle)
+            if self.viewModel.isLoaded {
+                List {
 
-                        if let subscriptionInformation = self.viewModel.subscriptionInformation {
+                    if let subscriptionInformation = self.viewModel.subscriptionInformation {
+                        Section {
                             SubscriptionDetailsView(
                                 subscriptionInformation: subscriptionInformation,
                                 localization: self.localization,
                                 refundRequestStatusMessage: self.viewModel.refundRequestStatusMessage)
                         }
+                    }
 
+                    Section {
                         ManageSubscriptionsButtonsView(viewModel: self.viewModel,
                                                        loadingPath: self.$viewModel.loadingPath)
-                    } else {
-                        ProgressView()
-                            .progressViewStyle(CircularProgressViewStyle())
+                    } header: {
+                        if let subtitle = self.viewModel.screen.subtitle {
+                            Text(subtitle)
+                                .textCase(nil)
+                        }
                     }
                 }
-                .padding([.horizontal, .bottom])
-                .frame(maxWidth: 400)
+            } else {
+                TintedProgressView()
             }
         }
         .toolbar {
@@ -123,7 +120,7 @@ struct ManageSubscriptionsView: View {
 
 }
 
-@available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
+@available(iOS 15.0, *)
 @available(macOS, unavailable)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
@@ -137,143 +134,7 @@ private extension ManageSubscriptionsView {
 
 }
 
-@available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
-@available(macOS, unavailable)
-@available(tvOS, unavailable)
-@available(watchOS, unavailable)
-struct SubtitleTextView: View {
-
-    private(set) var subtitle: String?
-    @Environment(\.appearance)
-    private var appearance: CustomerCenterConfigData.Appearance
-    @Environment(\.colorScheme)
-    private var colorScheme
-
-    var body: some View {
-        let textColor = Color.from(colorInformation: appearance.textColor, for: colorScheme)
-
-        if let subtitle {
-            Text(subtitle)
-                .font(.subheadline)
-                .padding([.horizontal])
-                .multilineTextAlignment(.center)
-                .applyIf(textColor != nil, apply: { $0.foregroundColor(textColor) })
-        }
-    }
-
-}
-
-@available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
-@available(macOS, unavailable)
-@available(tvOS, unavailable)
-@available(watchOS, unavailable)
-struct SubscriptionDetailsView: View {
-
-    let iconWidth = 22.0
-    let subscriptionInformation: SubscriptionInformation
-    let localization: CustomerCenterConfigData.Localization
-    let refundRequestStatusMessage: String?
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: 8) {
-            VStack(alignment: .leading) {
-                Text("\(subscriptionInformation.title)")
-                    .font(.headline)
-
-                let explanation = subscriptionInformation.active ? (
-                     subscriptionInformation.willRenew ?
-                            localization.commonLocalizedString(for: .subEarliestRenewal) :
-                            localization.commonLocalizedString(for: .subEarliestExpiration)
-                    ) : localization.commonLocalizedString(for: .subExpired)
-
-                Text("\(explanation)")
-                    .font(.subheadline)
-                    .foregroundColor(.secondary)
-                    .multilineTextAlignment(.leading)
-            }.padding([.bottom], 10)
-
-            Divider()
-                .padding(.bottom)
-
-            VStack(alignment: .leading, spacing: 16.0) {
-                HStack(alignment: .center) {
-                    Image(systemName: "coloncurrencysign.arrow.circlepath")
-                        .accessibilityHidden(true)
-                        .frame(width: iconWidth)
-                    VStack(alignment: .leading) {
-                        Text(localization.commonLocalizedString(for: .billingCycle))
-                            .font(.caption2)
-                            .foregroundColor(.secondary)
-                            .textCase(.uppercase)
-                        Text("\(subscriptionInformation.durationTitle)")
-                            .font(.body)
-                    }
-                }
-
-                HStack(alignment: .center) {
-                    Image(systemName: "coloncurrencysign")
-                        .accessibilityHidden(true)
-                        .frame(width: iconWidth)
-                    VStack(alignment: .leading) {
-                        Text(localization.commonLocalizedString(for: .currentPrice))
-                            .font(.caption2)
-                            .foregroundColor(.secondary)
-                            .textCase(.uppercase)
-                        Text("\(subscriptionInformation.price)")
-                            .font(.body)
-                    }
-                }
-
-                if let nextRenewal =  subscriptionInformation.expirationDateString {
-                    let expirationString = subscriptionInformation.active ? (
-                        subscriptionInformation.willRenew ?
-                            localization.commonLocalizedString(for: .nextBillingDate) :
-                            localization.commonLocalizedString(for: .expires)
-                    ) : localization.commonLocalizedString(for: .expired)
-
-                    HStack(alignment: .center) {
-                        Image(systemName: "calendar")
-                            .accessibilityHidden(true)
-                            .frame(width: iconWidth)
-                        VStack(alignment: .leading) {
-                            Text("\(expirationString)")
-                                .font(.caption2)
-                                .foregroundColor(.secondary)
-                                .textCase(.uppercase)
-                            Text("\(String(describing: nextRenewal))")
-                                .font(.body)
-                        }
-                    }
-                }
-
-                if let refundRequestStatusMessage = refundRequestStatusMessage {
-                    HStack(alignment: .center) {
-                        Image(systemName: "arrowshape.turn.up.backward")
-                            .accessibilityHidden(true)
-                            .frame(width: iconWidth)
-                        VStack(alignment: .leading) {
-                            Text(localization.commonLocalizedString(for: .refundStatus))
-                                .font(.caption2)
-                                .foregroundColor(.secondary)
-                                .textCase(.uppercase)
-                            Text("\(refundRequestStatusMessage)")
-                                .font(.body)
-                        }
-                    }
-                }
-            }
-        }
-        .padding(24.0)
-        .background(Color(UIColor.tertiarySystemBackground))
-        .cornerRadius(20)
-        .shadow(color: .black.opacity(0.1), radius: 10, x: 0.0, y: 10)
-        .padding([.top, .bottom])
-        .padding(.bottom)
-    }
-
-}
-
-@available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
+@available(iOS 15.0, *)
 @available(macOS, unavailable)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
@@ -290,23 +151,21 @@ struct ManageSubscriptionsButtonsView: View {
     private var localization: CustomerCenterConfigData.Localization
 
     var body: some View {
-        VStack(spacing: 16) {
-            let filteredPaths = self.viewModel.screen.paths.filter { path in
+        let filteredPaths = self.viewModel.screen.paths.filter { path in
 #if targetEnvironment(macCatalyst)
-                return path.type == .refundRequest
+            return path.type == .refundRequest
 #else
-                return true
+            return true
 #endif
-            }
-            ForEach(filteredPaths, id: \.id) { path in
-                ManageSubscriptionButton(path: path, viewModel: self.viewModel)
-            }
+        }
+        ForEach(filteredPaths, id: \.id) { path in
+            ManageSubscriptionButton(path: path, viewModel: self.viewModel)
         }
     }
 
 }
 
-@available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
+@available(iOS 15.0, *)
 @available(macOS, unavailable)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
@@ -327,7 +186,6 @@ struct ManageSubscriptionButton: View {
                 Text(path.title)
             }
         })
-        .buttonStyle(ProminentButtonStyle())
         .disabled(self.viewModel.loadingPath != nil)
         .restorePurchasesAlert(isPresented: self.$viewModel.showRestoreAlert)
         .sheet(item: self.$viewModel.promotionalOfferData,
@@ -346,52 +204,40 @@ struct ManageSubscriptionButton: View {
 }
 
 #if DEBUG
-@available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
+@available(iOS 15.0, *)
 @available(macOS, unavailable)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
 struct ManageSubscriptionsView_Previews: PreviewProvider {
 
     static var previews: some View {
-        let viewModelMonthlyRenewing = ManageSubscriptionsViewModel(
-            screen: CustomerCenterConfigTestData.customerCenterData.screens[.management]!,
-            subscriptionInformation: CustomerCenterConfigTestData.subscriptionInformationMonthlyRenewing,
-            customerCenterActionHandler: nil,
-            refundRequestStatusMessage: "Refund granted successfully!")
-        ManageSubscriptionsView(viewModel: viewModelMonthlyRenewing)
-            .previewDisplayName("Monthly renewing")
-            .environment(\.localization, CustomerCenterConfigTestData.customerCenterData.localization)
-            .environment(\.appearance, CustomerCenterConfigTestData.customerCenterData.appearance)
 
-        let viewModelYearlyExpiring = ManageSubscriptionsViewModel(
-            screen: CustomerCenterConfigTestData.customerCenterData.screens[.management]!,
-            subscriptionInformation: CustomerCenterConfigTestData.subscriptionInformationYearlyExpiring,
-            customerCenterActionHandler: nil)
-        ManageSubscriptionsView(viewModel: viewModelYearlyExpiring)
-            .previewDisplayName("Yearly expiring")
-            .environment(\.localization, CustomerCenterConfigTestData.customerCenterData.localization)
-            .environment(\.appearance, CustomerCenterConfigTestData.customerCenterData.appearance)
+        CompatibilityNavigationStack {
+            let viewModelMonthlyRenewing = ManageSubscriptionsViewModel(
+                screen: CustomerCenterConfigTestData.customerCenterData.screens[.management]!,
+                subscriptionInformation: CustomerCenterConfigTestData.subscriptionInformationMonthlyRenewing,
+                customerCenterActionHandler: nil,
+                refundRequestStatusMessage: "Refund granted successfully!")
+            ManageSubscriptionsView(viewModel: viewModelMonthlyRenewing)
+                .previewDisplayName("Monthly renewing")
+                .environment(\.localization, CustomerCenterConfigTestData.customerCenterData.localization)
+                .environment(\.appearance, CustomerCenterConfigTestData.customerCenterData.appearance)
+        }
+
+        CompatibilityNavigationStack {
+            let viewModelYearlyExpiring = ManageSubscriptionsViewModel(
+                screen: CustomerCenterConfigTestData.customerCenterData.screens[.management]!,
+                subscriptionInformation: CustomerCenterConfigTestData.subscriptionInformationYearlyExpiring,
+                customerCenterActionHandler: nil)
+            ManageSubscriptionsView(viewModel: viewModelYearlyExpiring)
+                .previewDisplayName("Yearly expiring")
+                .environment(\.localization, CustomerCenterConfigTestData.customerCenterData.localization)
+                .environment(\.appearance, CustomerCenterConfigTestData.customerCenterData.appearance)
+        }
     }
 
 }
 
-@available(iOS 15.0, macOS 12.0, tvOS 15.0, watchOS 8.0, *)
-@available(macOS, unavailable)
-@available(tvOS, unavailable)
-@available(watchOS, unavailable)
-struct SubscriptionDetailsView_Previews: PreviewProvider {
-
-    static var previews: some View {
-        SubscriptionDetailsView(
-            subscriptionInformation: CustomerCenterConfigTestData.subscriptionInformationMonthlyRenewing,
-            localization: CustomerCenterConfigTestData.customerCenterData.localization,
-            refundRequestStatusMessage: "Success"
-        )
-        .previewDisplayName("Subscription Details - Monthly")
-        .padding()
-    }
-
-}
 #endif
 
 #endif
